test(options): shard oversized option tests
